Sagas directory structured.
