+++ src/actions/themeActions.js
@@ -1,8 +1,8 @@
-export const loadedAppThemeSuccessAction = appTheme => ({
-  type: 'LOADED_APP_THEME_SUCCESS_ACTION',
+export const loadAppThemeSuccessAction = appTheme => ({
+  type: 'LOAD_APP_THEME_SUCCESS_ACTION',
   appTheme,
 });
 
 export const loadAppThemeFailedAction = () => ({
-  type: 'LOADED_APP_THEME_FAILED_ACTION',
+  type: 'LOAD_APP_THEME_FAILED_ACTION',
 });
