--- src/sagas/assetFilesSaga.js
@@ -1,64 +1,0 @@
-import { put } from 'redux-saga/effects';
-
-import {
-  fetchFileConfigFailedAction,
-  fetchFileConfigSuccessAction,
-  loadedAppDataSuccessAction,
-  loadAppDataFailedAction,
-  loadBusCoordinatorsDataSuccessAction,
-  loadBusCoordinatorsDataFailedAction,
-} from 'actions/assetFilesActions';
-import {
-  getAppConfig,
-  getBusCoordinatorsConfig,
-} from 'apis/core';
-import {
-  fetchFileConfig,
-} from 'apis/file';
-import { setAppConfigAction } from 'actions/app';
-
-export function* fetchFilesConfigSaga() {
-  const errorMessage = 'Unable to fetch file config.';
-  try {
-    const response = yield fetchFileConfig();
-    if (response) {
-      yield put(fetchFileConfigSuccessAction(response));
-    } else {
-      yield put(fetchFileConfigFailedAction(errorMessage));
-    }
-  } catch (e) {
-    console.error(e);
-    yield put(fetchFileConfigFailedAction(errorMessage));
-  }
-}
-
-export function* getAppConfigSaga() {
-  const errorMessage = 'Unable to fetch config.';
-  try {
-    const appConfig = yield getAppConfig();
-    if (appConfig) {
-      yield put(loadedAppDataSuccessAction(appConfig));
-      yield put(setAppConfigAction({ ...appConfig }));
-    } else {
-      yield put(loadAppDataFailedAction(errorMessage));
-    }
-  } catch (e) {
-    console.error(e);
-    yield put(loadAppDataFailedAction(errorMessage));
-  }
-}
-
-export function* getBusCoordinatorsConfigSaga() {
-  const errorMessage = 'Unable to fetch  bus coordinators config.';
-  try {
-    const response = yield getBusCoordinatorsConfig();
-    if (response) {
-      yield put(loadBusCoordinatorsDataSuccessAction(response));
-    } else {
-      yield put(loadBusCoordinatorsDataFailedAction(errorMessage));
-    }
-  } catch (e) {
-    console.error(e);
-    yield put(loadBusCoordinatorsDataFailedAction(errorMessage));
-  }
-}
